Initialize symlink policy variable in directory traversal.
Currently, chromiumos_get_symlink_traversal_policy() assumes that
either the dentry passed in or some directory higher up the path will
contain metadata regarding the symlink traversal policy. If this is
not the case, the function will return the uninitialized "policy"
variable.

This CL initializes that variable to
CHROMIUMOS_SYMLINK_TRAVERSAL_INHERIT, so the function still returns the
correct value even if no directory on the path has the metadata.

Currently, security/chromiumos/lsm.c is the only file that calls this
function, and in that case the uninitialized variable would only cause
unexpected behavior if no directory on the path has the metadata and
the uninitialized variable happens to equal
CHROMIUMOS_SYMLINK_TRAVERSAL_BLOCK.

BUG=chromium:655606
TEST=ran autotest client/site_tests/security_StatefulSymlinkTraversal

Signed-off-by: Micah Morton <mortonm@chromium.org>

Change-Id: I5ddcdd27298a0b7f74593aa1b303572803213c85
Reviewed-on: https://chromium-review.googlesource.com/792095
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>