CHROMIUM: arm64: dts: rockchip: convert Scarlet rev0 to use of-graph
We will be converting the driver over to use a common synopsys bridge
driver, and that plays nicely with us when we have the right
'remote-endpoint' links. The rockchip binding already documents this
usage but the driver doesn't enforce it; however, the new driver will,
so let's get this correct now.

Along the way, let's give the 'panel' a proper unit address, since it
has a 'reg' property. Also drop the unnecessary 'status = "okay"'.

BUG=b:65001779
TEST=build and boot scarlet rev0, display still works

Change-Id: I3f096bb37707aab5915cb006be2aa4e020adc8ab
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/781380
Reviewed-by: Douglas Anderson <dianders@chromium.org>