BACKPORT: drm/i915: Apply Display WA #1183 on skl, kbl, and cfl
Display WA #1183 was recently added to workaround
"Failures when enabling DPLL0 with eDP link rate 2.16
or 4.32 GHz and CD clock frequency 308.57 or 617.14 MHz
(CDCLK_CTL CD Frequency Select 10b or 11b) used in this
 enabling or in previous enabling."

This workaround was designed to minimize the impact only
to save the bad case with that link rates. But HW engineers
indicated that it should be safe to apply broadly, although
they were expecting the DPLL0 link rate to be unchanged on
runtime.

We need to cover 2 cases: when we are in fact enabling DPLL0
and when we are just changing the frequency with small
differences.

This is based on previous patch by Rodrigo Vivi with suggestions
from Ville Syrjl.

Conflicts:
1. Fix context different in skl_set_cdclk() function.
2. Replace IS_GEN9_BC as (IS_SKYLAKE || IS_KABYLAKE).
3. No CFL supported in v4.4.

BUG=b:74003991
TEST= suspend_stress_test -c 500 without "DPLL0 not locked" warning.

Cc: Arthur J Runyan <arthur.j.runyan@intel.com>
Cc: Ville Syrjl <ville.syrjala@linux.intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Cc: stable@vger.kernel.org
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20171204232210.4958-1-lucas.demarchi@intel.com
(cherry picked from commit 53421c2fe99ce16838639ad89d772d914a119a49)
Signed-off-by: Vincente Tsou <vincente.tsou@intel.com>

Change-Id: I9dbee07915423a400df082f7c423e3d41300ff06
Reviewed-on: https://chromium-review.googlesource.com/967643
Commit-Ready: Vincente Tsou <vincente.tsou@intel.com>
Tested-by: Vincente Tsou <vincente.tsou@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>