BACKPORT: FROMLIST: HID: input: do not increment usages when a duplicate is found
This is something that bothered us from a long time. When hid-input
doesn't know how to map a usage, it uses *_MISC. But there is something
else which increments the usage if the evdev code is already used.

This leads to few issues:
- some devices may have their ABS_X mapped to ABS_Y if they export a bad
  set of usages (see the DragonRise joysticks IIRC -> fixed in a specific
  HID driver)
- *_MISC + N might (will) conflict with other defined axes (my Logitech
  H800 exports some multitouch axes because of that)
- this prevents to freely add some new evdev usages, because "hey, my
  headset will now report ABS_COFFEE, and it's not coffee capable".

So let's try to kill this nonsense, and hope we won't break too many
devices.

I my headset case, the ABS_MISC axes are created because of some
proprietary usages, so we might not break that many devices.

For backward compatibility, a quirk HID_QUIRK_INCREMENT_USAGE_ON_DUPLICATE
is created and can be applied to any device that needs this behavior.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

BUG=chromium:825288
TEST=Build on guado, try connecting MeetUp device and verify it does not
list any extra absolute axes.

(am from https://patchwork.kernel.org/patch/10102635/ and adjust hid.h
to not use BIT() macros but plain constants)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: Id86933886d98ea4f5d7bb27784876bc7e56501c2
Reviewed-on: https://chromium-review.googlesource.com/1017864
Reviewed-by: Benson Leung <bleung@chromium.org>