CHROMIUM: iwl7000: pcie: remove non-responsive device
If we fail to to grab NIC access because the device is not responding
(i.e. CSR_GP_CNTRL returns 0xFFFFFFFF), remove the device from the PCI
bus, to avoid any further damage, and to let the user space rescan.

This aligns Rajat's patch that was applied to chromeos-3.18 with what
we had in our tree.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Rajat Jain <rajatja@google.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: b3b0fb46f35f3a7399ccd433417ab87b2df50cb5