BACKPORT: KVM: vmx: shadow more fields that are read/written on every vmexits
Compared to when VMCS shadowing was added to KVM, we are reading/writing
a few more fields: the PML index, the interrupt status and the preemption
timer value.  The first two are because we are exposing more features
to nested guests, the preemption timer is simply because we have grown
a new optimization.  Adding them to the shadow VMCS field lists reduces
the cost of a vmexit by about 1000 clock cycles for each field that exists
on bare metal.

On the other hand, the guest BNDCFGS and TSC offset are not written on
fast paths, so remove them.

Suggested-by: Jim Mattson <jmattson@google.com>
Cc: Jim Mattson <jmattson@google.com>
Cc: Wanpeng Li <wanpeng.li@hotmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>

BUG=b:120040011
TEST=Built

(cherry picked from commit c5d167b27e00026711ad19a33a23d5d3d562148a)
Signed-off-by: Matt Delco <delco@chromium.org>
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: Ica723cef00f0fccafbbf2018e75d127e04df7a21
Reviewed-on: https://chromium-review.googlesource.com/1766535
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>