CHROMIUM: oom: Chill out and don't print refusals to die so often
As of the commit ("CHROMIUM: oom: do not select new victim if current
one is runnable") we may print out about a task refusing to die but
then not select a new victim.  In that case then each subsequent call
into the OOM call would print another new refusal.

Let's avoid that by pretending that we selected a victim if we noticed
an existing victim that is runnable.

BUG=chromium:795344
TEST='mmm_donut.py -n200 -s10 -z50' on kevin w/ no ui, no slub debug

Change-Id: Id628c601483951adc54b7e4358512912bf96fbb9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/829921
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>