CHROMIUM: oom: Report info about the killer in OOM kills
Report which task (name/PID) is doing the killing when an OOM kill
happens.  Also report nvcsw (voluntary switches) and nivcsw
(involuntary switches) for the killer.  The idea here is that if a
task is runnable but refusing to die that perhaps the killer is just
looping too much.

BUG=chromium:795344
TEST='mmm_donut.py -n200 -s10 -z50' on kevin w/ no ui, no slub debug

Change-Id: I702a28e8ff5c5feb177f81be2a39f29c251e8890
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/831054
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 5f420c83308f9f97b81494a6df192f11a1082b73)