CHROMIUM: media: cio2: adjust watermark value to address the frame corruption issue
Root cause is bus/DMA throughput. It could work by tuning this
watermark value. in idle mode the system bus (as attached) will be
driven slower for lower power mode. And since the camera data will need
go through bus and DMA to be stored to DDR, so the bus might be
overflowed then the full data could not be kept. (the valid amount of
data in each corrupted image is also varied. )

BUG=b:72675688
TEST=Test by camera APP for stable preview on DUT

Change-Id: If2acfce5227a4b86f34f57edfb434dd0dd377582
Signed-off-by: Andy Yeh <andy.yeh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/835969
Reviewed-by: Tomasz Figa <tfiga@chromium.org>