CHROMIUM: iwl7000: pcie: get a reference before rescannning device
There is a potential race-condition in the PCI device rescan code
because the device may be removed by someone else while we have the
work scheduled, so when the work runs, there is no guarantee that our
reference is still valid.

To solve this, take a pci device reference by calling pci_dev_get()
and release it when the work is done.  This will prevent the reference
from being invalid when the work runs.

Additionally, add the PCI register dump back in the case we got
0xffffffff and save the parent before removing the device, so we can
still reference it in the rescan.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: faae88cd7bf314b6217da0617d9a936579b553ed