BACKPORT: tpm: delete the TPM_TIS_CLK_ENABLE flag
This flag is only used to warn if CLKRUN_EN wasn't disabled on Braswell
systems, but the only way this can happen is if the code is not correct.

So it's an unnecessary check that just makes the code harder to read.

BUG=chromium:944292
TEST=Build and run on cyan

Suggested-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Reviewed-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Tested-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
(cherry picked from commit 764325add6c2ad1641199edde7b2995495fc2d7c)

[backport: File name changes, enum tpm_tis_flags needs at least one flag]
Signed-off-by: Guenter Roeck <groeck@chromium.org>

Change-Id: I84bcfac2bdccb44d5c50610d9fd4b04db8f1bb07
Reviewed-on: https://chromium-review.googlesource.com/1533064
Commit-Ready: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>