UPSTREAM: tpm: only attempt to disable the LPC CLKRUN if is already enabled
Commit 5e572cab92f0 ("tpm: Enable CLKRUN protocol for Braswell systems")
added logic in the TPM TIS driver to disable the Low Pin Count CLKRUN
signal during TPM transactions.

Unfortunately this breaks other devices that are attached to the LPC bus
like for example PS/2 mouse and keyboards.

One flaw with the logic is that it assumes that the CLKRUN is always
enabled, and so it unconditionally enables it after a TPM transaction.

But it could be that the CLKRUN# signal was already disabled in the LPC
bus and so after the driver probes, CLKRUN_EN will remain enabled which
may break other devices that are attached to the LPC bus but don't have
support for the CLKRUN protocol.

Fixes: 5e572cab92f0 ("tpm: Enable CLKRUN protocol for Braswell systems")
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Tested-by: James Ettle <james@ettle.org.uk>

---

This patch fixes the bug reported for the Fedora kernel [0] and the kernel
bugzilla [1]. The issue and the propossed solution were discussed in this
[2] thread.

[0]: https://bugzilla.redhat.com/show_bug.cgi?id=1498987
[1]: https://bugzilla.kernel.org/show_bug.cgi?id=197287
[2]: https://patchwork.kernel.org/patch/10119417/
(cherry picked from commit 6c9f0ce0dffe64da2204f38b0fd90f3ae2a8903c)

Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>

BUG=b:68224689
TEST=suspend_stress_test --memory_check passed 10000 iterations on cyan

Change-Id: I977eda80876814ddaa0b9537f0f50e430f0c15e7
Reviewed-on: https://chromium-review.googlesource.com/853634
Commit-Ready: Azhar Shaikh <azhar.shaikh@intel.com>
Tested-by: Jeremy Lin <jeremy.p.lin@intel.com>
Reviewed-by: Jack Lai <jack.lai@intel.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>