CHROMIUM: iommu/rockchip: Only count for enabled IOMMUs
Currently we are counting for all IOMMUs in rk_iommu_of_setup() and
set bus IOMMU after all of them been probed.

This would not work if any of the IOMMUs are disabled, which means
their probe would never be called.

So only count for enabled IOMMUs to avoid that.

Fixes: c0346906a871 ("CHROMIUM: iommu/rockchip: Set bus IOMMU only after all IOMMUs probe")

BUG=b:36227021
TEST=Test on scarlet, can boot to UI with a disabled IOMMU
TEST=Bob can display normally with the whole ISP driver

Change-Id: Icb7283baad753d4083a6c7d35cf4b8579af36951
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Signed-off-by: Shunqian Zheng <zhengsq@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/848572
Reviewed-by: Ricky Liang <jcliang@chromium.org>