BACKPORT: ANDROID: binder: remove waitqueue when thread exits.
(cherry picked from commit f5cb779ba163)

binder_poll() passes the thread->wait waitqueue that
can be slept on for work. When a thread that uses
epoll explicitly exits using BINDER_THREAD_EXIT,
the waitqueue is freed, but it is never removed
from the corresponding epoll data structure. When
the process subsequently exits, the epoll cleanup
code tries to access the waitlist, which results in
a use-after-free.

Prevent this by using POLLFREE when the thread exits.

BUG=b:142155285
TEST=See bug.

Signed-off-by: Martijn Coenen <maco@android.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Mattias Nissler <mnissler@chromium.org>

Change-Id: I7fece7fcb80504ccbb4944829d9bfbcf498fbdb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1843031
Reviewed-by: Greg Kroah-Hartman <gregkh@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>