FROMLIST: bluetooth: xbox controller bug fix
The current L2CAP implementation does not change any options if the other
side respons with "unknown options", but does if "unaccepted options" is
the answer. It is up to the implementation to decide on the effort spent
on config negotiations, therefore the current implementation is correct
at this point - but many devices (like Xbox One S controllers) are not
useable this way. A workaround for many users therefore is to
disable_ertm, since this is often the option which is unknown. I would
prefer to try it again with altered options instead of globally disable
ERTM. In result, I suggest the following patch. It simply adds a new case
(L2CAP_CONF_UNKNOWN), which does nothing but falling through to
L2CAP_CONF_UNACCEPT.

Signed-off-by: Florian Dollinger <dollinger.florian@gmx.de>

(am from https://lore.kernel.org/linux-bluetooth/fcdc7234-627c-dff1-d702-c574c975be54@gmx.de/)

BUG=chromium:930997
TEST=verify that pairing with an Xbox controller works

Change-Id: Id4403a87f7abcb1c82148eb4000005cbeb80dc49
Signed-off-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1757593
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>