CHROMIUM: add case of CONFIG_ARM64 in drm_clflush_sg
The dma API apparent works for ARM64. We thought this might not
be the case, since we were under the impression there's a swiotlb
backend for the API (like Intel has).  Atleast for v3.18, this
works.

Let's add a case for it in drm_cache file.

BUG=chromium:654131
TEST=vgem_test works on Elm

Change-Id: I2b385204dd901ef3261e3dc263e3402285e18abf
Reviewed-on: https://chromium-review.googlesource.com/860737
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>