CHROMIUM: drm/vgem: flush after getting pages
The pages that are fetched in vgem_get_pages() may be still be in
the cache. We're experiencing a case where eviction of the cache
races with the writes we do. To fix this, flush the cache during
vgem_get_pages().

BUG=chromium:654131
TEST=vgem_test works on Elm

Change-Id: I00c99f63e10aa9001043ce74912297e2bffb6478
Reviewed-on: https://chromium-review.googlesource.com/860738
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>