HACK: drm/panel: Innolux P097PFG: Add initialization sequence
Innolux didn't deliver a working init sequence yet for devices without
OTP programming. The sequence in this change has been derived from a
register dump of a mostly working panel with OTP. It is not meant to
be final, but to make devices with unprogrammed OTP work, while Innolux
is figuring out a proper sequence. There is a known issue with an
artifact line in the lower third of the display.

BUG=b:69689064, b:72191820
TEST=boot on dru with an Innolux panel with unprogrammed OTP memory
  => the panel shows an odd pattern shortly after boot, but generally
    works after booting has completed.
  repeat 10x

Change-Id: Idc8d8920523314408cd7d686e040e76cc12c1a1e
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/872432
Reviewed-by: Brian Norris <briannorris@chromium.org>
(cherry picked from commit 538e51760666ccbb21689d48f4a490f6b9537f76)
Reviewed-on: https://chromium-review.googlesource.com/881741
Tested-by: Brian Norris <briannorris@chromium.org>