HACK: drm/panel: Innolux P097PFG: Initialize twice at boot
For unknown reasons panels without OTP programming are often not properly
initialized after the first invocation of innolux_panel_prepare(). A
re-initialization appears to initialize the panel reliably. Obviously
this is not a long-term solution, but let's use this hack for now to
make the panels usable at all, while the root cause is being investigated.

BUG=b:69689064, b:72191820
TEST=see CL:872432

Change-Id: Ib4aed33962c2908fd793b9327d5cd7579cdf459e
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/876944
Reviewed-by: Brian Norris <briannorris@chromium.org>