CHROMIUM: Fix warnings in Rockchip components.
This change makes some no-op modifications in order to enable Werror on
the 3.14 kernel. Most of the warnings here are casts between integers and
pointers of different sizes. These casts would be problematic if the
physical addresses used here were above the 4GB line and were being cast
to a 32-bit virtual address. Because this is non-upstreamed
Rockchip-specific code, these physical addresses are known to be in the
first 4GB.

There's also a deletion of an unused local variable in here.

BUG=chromium:702788
TEST=Build and boot veyron_minnie.

Change-Id: I9ca72c1700748089540bb7c6170872cf7f5c4795
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/877331
Commit-Ready: Evan Green <evgreen@google.com>
Tested-by: Evan Green <evgreen@google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>