CHROMIUM: drm/i915: Wait for vblank after disabling cursor
The cursor disable write is latched on vsync but if we do a modeset
before that happens the write seems to be discarded and the cursor
stays on. The CURCNTL register will still read back 0 as if the write
took, but underlying hardware didn't change. We don't disable the
cursor often, so we can just force the pre-GEN9 path for cursor
that waits for vblank.

BUG=b:70546200, b:70534167
TEST=Toggle primary/extended display as described in bug

Change-Id: Ia9f8bddfe5c3b65ebe457dff1755547751ef0f47
Signed-off-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/877396
Reviewed-by: Sean Paul <seanpaul@google.com>
(cherry picked from commit 24437c01ae51d2f3eaa301a8c901e039a5646c86)
Reviewed-on: https://chromium-review.googlesource.com/903528
Tested-by: Sean FS Chiang <sean_fs_chiang@wistron.corp-partner.google.com>
Commit-Queue: Sean FS Chiang <sean_fs_chiang@wistron.corp-partner.google.com>