CHROMIUM: config: Add chromiumos-qualcomm config
This change adds a chromiumos-qualcomm.flavour.config. It would
contain the minimal set of configs needed to boot the SDM845 MTP
to RAM disk, except that it was normalized. So the needed configs
that are not yet in this tree got obliterated. That was done so that
this change could land while still having "normalize config" be a
no-op.

I'll put a second change on top of this one in a moment that enables
the missing features, though that part cannot land until the features
they enable do.

BUG=chromium:795946
TEST=Build Rajendra's tree and boot cheza to RAM disk.

Change-Id: I5de6ee5909b053b7b42578a47262f6cea6d898f4
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/887920
Commit-Ready: Evan Green <evgreen@google.com>
Tested-by: Evan Green <evgreen@google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit de49895a65ba5a9c247fabc84573b25a70515658)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>