UPSTREAM: regulator: add PM suspend and resume hooks
In this patch, consumers are allowed to set suspend voltage, and this
actually just set the "uV" in constraint::regulator_state, when the
regulator_suspend_late() was called by PM core through callback when
the system is entering into suspend, the regulator device would act
suspend activity then.

And it assumes that if any consumer set suspend voltage, the regulator
device should be enabled in the suspend state.  And if the suspend
voltage of a regulator device for all consumers was set zero, the
regulator device would be off in the suspend state.

This patch also provides a new function hook to regulator devices for
resuming from suspend states.

Signed-off-by: Chunyan Zhang <zhang.chunyan@linaro.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit f7efad10b5c492892b1e5decf5d3ebb29fa5c9af)

BUG=b:74988960
TEST=Later patches pick clean; this series might be useful later

Change-Id: I38eb31dbaff2facb1d562b7a13304bc9e6b99122
Signed-off-by: Douglas Anderson <dianders@chromium.org>