CHROMIUM: sched: EAS: Fix build warnings
Recent versions of gcc report:

include/linux/sched.h:1172:42: warning:
	type qualifiers ignored on function return type

A pointer returned by a function is always assigned to something,
so it doesn't help to declare it const.

On top of that, 0day reports:

kernel/sched/fair.c:4938:34: error: duplicate 'const' declaration specifier
kernel/sched/sched.h:920:34: error: duplicate 'const' declaration specifier

One of the 'const' declarations in "const <type> const" is useless since
const is evaluated right to left. Presumably the idea is to declare both
the pointer and the data it points to as constant, so let's do that.

BUG=None
TEST=Build with changes applied

Change-Id: I22c155527801c827db5df5bb777d586ff592ae54
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/894611
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>