BACKPORT: arm64: entry: Apply BP hardening for suspicious interrupts from EL0
commit 30d88c0e3ace upstream.

It is possible to take an IRQ from EL0 following a branch to a kernel
address in such a way that the IRQ is prioritised over the instruction
abort. Whilst an attacker would need to get the stars to align here,
it might be sufficient with enough calibration so perform BP hardening
in the rare case that we see a kernel address in the ELR when handling
an IRQ from EL0.

BUG=b:72869496
TEST=build/boot on kevin

Change-Id: Id3f27c009cc511a6b7941e75c0109d8e41b63276
Reported-by: Dan Hettena <dhettena@nvidia.com>
Reviewed-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Alex Shi <alex.shi@linaro.org>
(cherry picked from commit 069dfbc5498eec2a987e6c6b689052f49e343796
 in git.linaro.org/kernel/linux-linaro-stable.git v4.4-spectre)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/985009