FROMLIST: irqchip/gic-v3-its: add ability to resend MAPC on resume
This adds functionality to resend the MAPC command to an ITS node on
resume. If the ITS is powered down during suspend and the collections
are not backed by memory, the ITS will lose that state. This just sets
up the known state for the collections after the ITS is restored.

This is enabled via the reset-on-suspend flag in the DTS for an ITS
that has a non-zero number of collections stored in it.

(am from https://patchwork.kernel.org/patch/10206511/)

Conflicts:
drivers/irqchip/irq-gic-v3-its.c
-Some functions not added yet
-Needed to use of node instead of fwnode_handle

BUG=b:62644399
TEST=suspend/resume on scarlet

Change-Id: I5dadb6b51c054d5057aacbc40be8ec54b008fa3f
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/890225
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>