CHROMIUM: config: Normalize splitconfig
This is just running:
  ./chromeos/scripts/kernelconfig oldconfig

...and then holding down the return key to accept the default for
everything.  It's official a no-op change (since this is what the
builder will do anyway when actually building the kernel), but to
justify:

CONFIG_KAISER is gone with and renamed:
  d441d63a125f Revert "BACKPORT: KAISER: Kernel Address Isolation"
  3e1457d6bf26 KPTI: Rename to PAGE_TABLE_ISOLATION

CONFIG_PAGE_TABLE_ISOLATION is default:
  3e1457d6bf26 KPTI: Rename to PAGE_TABLE_ISOLATION

CONFIG_RETPOLINE is default with:
  3c5e10905263 x86/retpoline: Add initial retpoline support

CONFIG_GENERIC_CPU_VULNERABILITIES is automatically selected when
needed:
  72cf81e43ba4 x86/cpu: Implement CPU vulnerabilites sysfs functions

The config changes for CONFIG_VIDEO_DW9807 and CONFIG_VIDEO_IMX258
where not well normalized in
  1459c4ff89ef CHROMIUM: config: Enable IMX258 camera and DW9807 VCM
...and the configs are newly available as of:
  03c3d1855a26 FROMLIST: media: dw9807: Add dw9807 vcm driver
  032efd70138a FROMLIST: media: imx258: Add imx258 camera sensor driver
...which explains why the rockchip 64 config didn't previously have
these listed as unset.  It seems sane not to enable them there.

BUG=None
TEST=None

Change-Id: Ie4b3e6f66d84a26a0e48277479c29cfe7bb6c965
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/906921
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>