CHROMIUM: iwl7000: Revert "iwlwifi: mvm: warn if seq_ctrl gets out of sync"
This patch caused much trouble and must be reverted.
It is the right thing to do to warn when we are out of sync,
but for now, it is causing too much noise.

We will need to debug this later though.

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 2be0785b1013d4a90663e34049580a9b73027933