UPSTREAM: Bluetooth: btrtl: Add RTL8723D and RTL8821C devices
The Bluetooth parts of RTL8723D and RTL8723B share the same lmp
subversion, thus we need to check both lmp subversion and hci revision
to distinguish the two. The same situation is true for RTL8821A and
RTL8821C. Accordingly, the selection code is revised.

To improve maintainability, a new id_table struct is defined, and an
array of such structs is constructed. Adding a new device can thus be
as simple as adding another value to the table.

Signed-off-by: Alex Lu <alex_lu@realsil.com.cn>
Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 907f84990924bf3a8d248c040dabeb5127ae6938)

BUG=b:139382142
TEST=use RTL8723D and RTL8723B to verify BT bring-up correctly.

Signed-off-by: Max Chou <max.chou@realtek.com>

Change-Id: I55030767c0e0c56864a4517df4c3fb921a44183f
Signed-off-by: Michael Sun <michaelfsun@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1775544
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Yoni Shavit <yshavit@chromium.org>