FROMLIST: virgl: fix caps ioctl size check.
The size check seems backwards, if userspace gives us a size
larger it is probably asking for a newer version of the caps,
so we just fill in the caps version we have at the smaller
size, and it can work it out from there.

Userspace will probably still need a fallback as old
kernel are out there.

Change-Id: I05136f3109f16fd2176c75140e28d50c4fa73ff5
Cc: <stable@vger.kernel.org>
Signed-off-by: Dave Airlie <airlied@redhat.com>
Reviewed-by: Stphane Marchesin <marcheu@chromium.org>
Signed-off-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/933727
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>