CHROMIUM: Revert "Bluetooth: btusb: Restore QCA Rome suspend/resume fix with a "rewritten" version"
This reverts commit 32a3eda07cc980af4758ff40b7ae9cf262168e50.

As per commit 629db59a0f14 ("CHROMIUM: Revert "UPSTREAM: Bluetooth: btusb:
driver to enable the usb-wakeup feature"") in chromeos-4.4, this patch results
in inband communication from the btusb device to cause the system to wake up.
While the hereby reverted patch is an attempt to fix that problem, it fails to
do so. Revert it prior to the merge of v4.4.116 into chromeos-4.4.

Change-Id: I679f380280621bd0235aa3a680bb3df0a561d57a
Signed-off-by: Guenter Roeck <groeck@chromium.org>