CHROMIUM: iwl7000: add shared clock PHY config flag for some devices
Some devices use a shared clock which is very sensitive to variations
and cause trouble in some situations.  We need to set a bit in the phy
configuration to indicate that to the FW.  To make this generic, add a
extra_phy_config_flags element to the device configuration and OR it
into the phy_cfg before sending it to the firmware.  And also create a
set of configurations for devices that use shared clocks and need this
extra bit to be set.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: e5d3d5180bf96d4ded0d7c871f6da49603a0b1cc