CHROMIUM: iwl7000: rx: unionize the RX descriptors
Most of the fields we care about are in the same location
in both versions.

Given that, the casting to different types is not only
error-prone, but redundant as well.

Unionize the RX descriptors.

Get rid of most of the if/else and of all the castings.

This also saves quite a few code lines.

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: cd34303236f05ebc332f1bed5819b1525447d72e