FROMLIST: rtc: cros-ec: return -ETIME when refused to set alarms in the past
Since accessing a Chrome OS EC based rtc is a slow operation, there is a
race window where if the alarm is set for the next second and the second
ticks over right before calculating the alarm offset.

In this case the current driver is setting a 0-second alarm, which would
be considered as disabling alarms by the EC(EC_RTC_ALARM_CLEAR).

This breaks, e.g., hwclock which relies on RTC_UIE_ON ->
rtc_update_irq_enable(), which sets a 1-second alarm and expects it to
fire an interrupt.

So return -ETIME when the alarm is in the past, follow __rtc_set_alarm().

Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
(am from https://patchwork.kernel.org/patch/10244153/)

BUG=chromium:815249, b:38310685
TEST=Test on scarlet:
     while hwclock -r;do sleep .1;done
     also, `test_that ... power_SuspendStress.bare`

Change-Id: I9b5d3ce17eed2181f346cf82c63d165bc8546708
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/936721
Commit-Ready: Jeffy Chen <jeffy.chen@rock-chips.com>