drm/rockchip: Stop using broken vop_win_id() helper
This is supposed to compute the window index, which is
win - vop->win (number of elements, not bytes), but then divides
that by sizeof(*win) and thus always returns 0.

The only place where we actually use this helper, we already have the
win_index in a local variable. Just use that instead and drop the
broken helper.

Change-Id: I22275e9852567febc6206d459d097e189fe4af39
Signed-off-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/942365
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Sean Paul <seanpaul@google.com>