CHROMIUM: drm/radeon: Increase threshold for stack size warning
With the default threshold of 2048 the following warning is raised
when the kernel is built with clang:

drivers/gpu/drm/radeon/radeon_combios.c:1452:6: error: stack frame size
  of 2560 bytes in function 'radeon_get_legacy_connector_info_from_table'
  [-Werror,-Wframe-larger-than=]
    bool radeon_get_legacy_connector_info_from_table(struct drm_device *dev)

clang is to blame here, with gcc the frame size is only 256 (more details
in chromium:817628)

This should be definitely be addressed in clang, but since warnings in
kernel builds are treated as errors we don't want to be blocked waiting
for a fix. As a stopgap increase the warning threshold to 4096 bytes
for the affected files. This change can be reverted once clang behaves
more civily.

BUG=chromium:817628
TEST=build for betty with clang

Change-Id: Id7885f37abd9aa7cb5e4183965b4fb5705eacbc5
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/944610
Reviewed-by: Guenter Roeck <groeck@chromium.org>