CHROMIUM: iio: Add SEMTECH SX9310 sensor driver
Add SEMTECH SX9310 driver.
The device has the following entry points:

Usual frequency:
- sampling_frequency
- sampling_frequency_available

ACPI UID:
- uid (can be used to distinguish multiple sensor instances)

Instant reading of current values for different sensors:
- in_proximity0_CS0_raw
- in_proximity1_CS1_raw
- in_proximity2_CS2_raw
- in_proximity3_CS3_raw
and associated events in events/

This driver expects the configuration of the device to be supplied via
ACPI, but provides a set of defaults for any/all missing settings.

Only ACPI - not device tree - is supported for now.

BUG=b:74363445
TEST=On Nocturne, check:
  - that the device is correctly probed and instantiated;
  - that it is possible to enable events on all channels and read from /dev

Change-Id: I0969fccdb6dec153f8fb8419d6f79cdf83b80df0
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1089826